# kloggerx-server/docs/swagger.yaml
basePath: /api/v1
definitions:
  model.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: KloggerX 知识协作平台 API 文档
  title: KloggerX API
  version: "1.0"
paths:
  /admin/departments:
    get:
      description: 获取部门树结构
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 获取部门树
      tags:
      - 系统管理
    post:
      consumes:
      - application/json
      description: 管理员创建新部门
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 创建部门
      tags:
      - 系统管理
  /admin/departments/{id}:
    delete:
      description: 管理员删除部门
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 删除部门
      tags:
      - 系统管理
    put:
      consumes:
      - application/json
      description: 管理员更新部门信息
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 更新部门
      tags:
      - 系统管理
  /admin/departments/{id}/members:
    get:
      description: 获取部门下的成员列表
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 获取部门成员
      tags:
      - 系统管理
  /admin/directories:
    get:
      description: 列出服务器目录结构
      parameters:
      - description: 目录路径
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 列出目录
      tags:
      - 系统管理
  /admin/feedback/reviews:
    get:
      description: 获取待审核的反馈列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 获取待审核列表
      tags:
      - 系统管理
  /admin/feedback/reviews/{id}/approve:
    post:
      consumes:
      - application/json
      description: 通过反馈审核
      parameters:
      - description: 审核ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 通过审核
      tags:
      - 系统管理
  /admin/feedback/reviews/{id}/reject:
    post:
      consumes:
      - application/json
      description: 拒绝反馈审核
      parameters:
      - description: 审核ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 拒绝审核
      tags:
      - 系统管理
  /admin/feedback/reviews/stats:
    get:
      description: 获取审核统计信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 审核统计
      tags:
      - 系统管理
...
  /admin/users:
    get:
      description: 管理员分页获取用户列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 管理员获取用户列表
      tags:
      - 系统管理
    post:
      consumes:
      - application/json
      description: 管理员创建新用户
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 管理员创建用户
      tags:
      - 系统管理
  /admin/users/{id}:
    delete:
      description: 管理员删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 管理员删除用户
      tags:
      - 系统管理
    put:
      consumes:
      - application/json
      description: 管理员更新用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 管理员更新用户
      tags:
      - 系统管理
  /admin/users/{id}/reset-password:
    post:
      consumes:
      - application/json
      description: 管理员重置用户密码
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - BearerAuth: []
      summary: 重置用户密码
      tags:
      - 系统管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"